Refactor
Extracted class for collecting and displaying warnings and error
messages.  Extracted method for figuring out which part of the line is
label, mnemonic, and argument.

More can be done.  Find out how to avoid passing line and number to each
method that reports errors.  Make classes for data and mnemonic info.