Improved graphical interface
Removed number pad and power button.

Implemented input stack.  Numbers are entered in a text field and then pushed onto a read-only stack.  Program takes input from the stack if available or waits until it's entered.

Output is accumulated in a read-only window.  It can be cleared with the "clear" button.

Trying to load a file that's not a program gives an error box.

The values of the program counter and input and output registers are displayed.