Fixed stepping
Added comments to LMC_Client.

Give step notification before execution.  Make sure notification is given for all steps.

Don't allow blocking on output.

Allow specifying the width of a Register entry field.  Provide a set() method that takes an integer and displays it with the proper number of leading zeros.  Show a green light in the input field when the program needs input.

Indicate the memory cell of the current operation with a filled progress bar.  Update the memory display on each step.

Show the accumulator register.